requirements.txt
requests==2.26.0
requests-toolbelt==0.9.1
yfinance
flask
python-dotenv
